+++ packages/react/src/components/Collapse/index.tsx
@@ -1,7 +1,7 @@
-import { ElementType, FC, useState } from 'react';
+import { ElementType, FC, useEffect, useState } from 'react';
 import { CaretDown, CaretUp } from '@phosphor-icons/react';
 
-import { CollapseContainerStyled, CollapseContentStyled, CollapseHeaderStyled } from './styles';
+import { CollapseContainerStyled, CollapseContentStyled, CollapseHeaderStyled, TitleHeaderStyled } from './styles';
 
 export interface ICollapseProps {
   title: string;
@@ -11,6 +11,8 @@ export interface ICollapseProps {
   color?: string
   colorIcon?: string
   fontSize?: 'xxs' | 'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2xl' | '4xl' | '5xl' | '6xl' | '7xl' | '8xl' | '9xl'
+  fontWeight?: 'regular' | 'medium' | 'semiBold' | 'bold'
+  isOpen?: boolean
 }
 
 export const Collapse: FC<ICollapseProps> = ({ 
@@ -20,19 +22,27 @@ export const Collapse: FC<ICollapseProps> = ({
   iconClose = CaretDown, 
   color = 'gray-800', 
   colorIcon = 'gray-800',
-  fontSize = 'md'
+  fontSize = 'md',
+  fontWeight = 'regular',
+  isOpen = false
 }) => {
-  const [isOpen, setIsOpen] = useState(false);
+  const [open, setOpen] = useState(false);
 
-  const Icon = isOpen ? iconOpen : iconClose;
+  const Icon = open ? iconOpen : iconClose;
+
+  useEffect(() => {
+    setOpen(isOpen)
+  }, [isOpen])
 
   return (
     <CollapseContainerStyled>
-      <CollapseHeaderStyled fontSize={fontSize} onClick={() => setIsOpen(!isOpen)}>
-        <span color={color}>{title}</span>
+      <CollapseHeaderStyled onClick={() => setOpen(!open)}>
+        <div style={{color}}>
+          <TitleHeaderStyled fontWeight={fontWeight} fontSize={fontSize}>{title}</TitleHeaderStyled>
+        </div>
         <Icon color={colorIcon} size={15} />
       </CollapseHeaderStyled>
-      <CollapseContentStyled isOpen={isOpen}>
+      <CollapseContentStyled isOpen={open}>
         {children}
       </CollapseContentStyled>
     </CollapseContainerStyled>